Add dedicated component
