+++ packages/web-pkg/src/components/SideBar/index.ts
@@ -3,3 +3,4 @@ export * from './types'
 
 export { default as CompareSaveDialog } from './CompareSaveDialog.vue'
 export { default as SideBar } from './SideBar.vue'
+export { default as WevDavDetails } from './WebDavDetails.vue'
